Refactor & remove filtering logic
